test: build connect and connack tool
