+++ tests/src/mqtt_build_tool/mod.rs
@@ -12,4 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod mqtt_build_tool;
+pub mod build_connack;
+pub mod build_connect;
